requirements.txt
Django==4.2
Pillow
psycopg2-binary
ipython
psycopg2

djangorestframework~=3.14.0